+++ Gu.Analyzers.Test/GU0012NullCheckParameterTests/Valid.cs
@@ -129,10 +129,10 @@ namespace N
 
         public C(string text)
         {
-#pragma warning disable GU0015 // Don't assign same more than once.
+#pragma warning disable GU0015 // Don't assign same more than once
             this.text = text ?? throw new ArgumentNullException(nameof(text));
             this.text = text;
-#pragma warning restore GU0015 // Don't assign same more than once.
+#pragma warning restore GU0015 // Don't assign same more than once
         }
     }
 }";