+++ ValidCode/Recursion.cs
@@ -213,7 +213,7 @@ namespace ValidCode
 
         internal void Meh()
         {
-#pragma warning disable IDE0079, GU0015 // Don't assign same more than once.
+#pragma warning disable IDE0079, GU0015 // Don't assign same more than once
             var value = this.RecursiveExpressionBodyProperty;
             value = this.RecursiveStatementBodyProperty;
             value = this.RecursiveExpressionBodyMethod();
@@ -227,7 +227,7 @@ namespace ValidCode
             RecursiveOut(string.Empty, out value);
             RecursiveRef(ref value);
             value = value;
-#pragma warning restore IDE0079, GU0015 // Don't assign same more than once.
+#pragma warning restore IDE0079, GU0015 // Don't assign same more than once
         }
 
         private static IDisposable RecursiveStatementBodyMethodWithOptionalParameter(IDisposable value, IEnumerable<IDisposable>? values = null)